LFF moved
My code is
constTag.innerHTML = "blah"
yet this server thought I wrote
constTag = "blah"

and because of that my server is blaming me for re-assigning the value of const

this is your falut, server. not me.

Oh also I moved LFF to another repository

And because of that I think commit count is useless from now on
so I'm going to not count this.